requirements.txt
python-dotenv==1.0.1
pydantic==2.6.1
PyYAML==6.0.1
numpy==1.26.4
pandas==2.2.0
sentence-transformers==2.6.0
faiss-cpu==1.7.4
spacy==3.7.4
rapidfuzz==3.6.1
tqdm==4.66.2
pytest==8.0.0
openai==1.12.0
jupyter>=1.0.0
ipykernel>=6.0.0
